testing to a config die on deflate or return with runtime UPGRADE
Third test .3 is a renamed zip file to force a deflate, DO NOT return running the Spartacus UPGRADEs, but KEEP other possible pending upgrade plugins remembered.
This latter seems not avoidable, while serendipity_plugin_api::setPluginInfo() will ALWAYS run to set versions and/or timestamp (mainly for the local).
If the UPGRADE runtime breaks with die(), the pluginlist table field 'upgrade_version' is reset to the current plugin version and Spartacus refuses to search for UPGRADES again until cache time is over.
So the best soution seems to nuke the local xml target in the deflate() to die() redirector.
We might think about having to additionally use clearstatcache().