adduser: Fix a very old bug preventing captchas to be actually build
This depends on the core spamblock plugin checking the timestamp, which can not be put into the $eventData stream via the independent adduser common.inc serendipity_common_adduser class.
Therefor the current fix is to send $serendipity['GET']['subpage'] = 'adduser'; and check this in spamblock right before the captcha generation.