social: Merge [s9y] Add per-entry social media image
event_social will use the first image from the entry for the OpenGraph
"og:image" meta property used by Facebook and Twitter; if there is none,
it will fall back to the default image from the plugin configuration.

The timeline and photo themes support setting a featured image that is used instead.

Add the possibility to set a social media image for each entry, regardless of theme.
If there is none set, fall back to timeline or photo featured images, or the first image
in the entry body, or the default image from the plugin configuration (in that particular order).